Update project files to include classes CustomStreamService and FlightInfo as seperate files. Correct  degree symbol display in WinUI. Update readme.metadata.json to reference new files. Remove unnecessary image reference from ArcGIS.WinUI.Viewer.csproj. Fix iOS CollectionView stuck when toggling between flight details
