+++ src/WPF/WPF.Viewer/Samples/Search/QueryDynamicEntities/QueryDynamicEntities.xaml.cs
@@ -19,19 +19,10 @@ using Esri.ArcGISRuntime.UI;
 using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
-using System.ComponentModel;
-using System.Diagnostics;
-using System.IO;
 using System.Linq;
-using System.Runtime.CompilerServices;
-using System.Text.Json;
-using System.Threading;
 using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Controls;
-using System.Windows.Markup;
-using System.Windows.Media;
-using System.Windows.Shapes;
 
 namespace ArcGIS.WPF.Samples.QueryDynamicEntities
 {
@@ -57,34 +48,42 @@ namespace ArcGIS.WPF.Samples.QueryDynamicEntities
             _ = Initialize();
         }
 
-        //Sets up the map, graphics overlay, and data source for flight tracking.
+        // Sets up the map, graphics overlay, and data source for flight tracking.
         private async Task Initialize()
         {
             try
             {
+                // Create the map with a basemap and initial viewpoint.
                 MyMapView.Map = new Map(BasemapStyle.ArcGISTopographic);
                 MyMapView.SetViewpoint(new Viewpoint(_phoenixAirport, 400000));
 
+                // Create a graphics overlay for the airport buffer.
                 _bufferGraphicsOverlay = new GraphicsOverlay();
                 MyMapView.GraphicsOverlays.Add(_bufferGraphicsOverlay);
 
+                // Create a 15-mile geodetic buffer around Phoenix Airport (PHX).
                 _phoenixAirportBuffer = GeometryEngine.BufferGeodetic(
                     _phoenixAirport, 15, LinearUnits.Miles, double.NaN, GeodeticCurveType.Geodesic);
 
+                // Create a semi-transparent fill symbol for the buffer and add it to the overlay.
                 var bufferSymbol = new SimpleFillSymbol(
                     SimpleFillSymbolStyle.Solid,
                     System.Drawing.Color.FromArgb(40, 255, 0, 0),
                     new SimpleLineSymbol(SimpleLineSymbolStyle.Solid, System.Drawing.Color.Black, 1));
 
+                // Add the buffer graphic to the overlay and hide it initially.
                 _bufferGraphicsOverlay.Graphics.Add(new Graphic(_phoenixAirportBuffer, bufferSymbol));
                 _bufferGraphicsOverlay.IsVisible = false;
 
+                // Initialize the collection for query results and bind it to the results list.
                 _queryResults = new ObservableCollection<FlightInfo>();
                 ResultsList.ItemsSource = _queryResults;
 
+                // Hide the results panel and flight number dialog initially.
                 ResultsPanel.Visibility = Visibility.Collapsed;
                 FlightNumberDialog.Visibility = Visibility.Collapsed;
 
+                // Set up the dynamic entity data source and layer.
                 await SetupDynamicEntityDataSource();
             }
             catch (Exception ex)
@@ -94,11 +93,12 @@ namespace ArcGIS.WPF.Samples.QueryDynamicEntities
             }
         }
 
-        // Creates and configures the dynamic entity data source and layer for real-time flight tracking
+        // Creates and configures the dynamic entity data source and layer for real-time flight tracking.
         private async Task SetupDynamicEntityDataSource()
         {
             try
             {
+                // Define the fields for the dynamic entity data source.
                 var fields = new List<Field>
                 {
                     new Field(FieldType.Text, "flight_number", "", 50),
@@ -115,9 +115,11 @@ namespace ArcGIS.WPF.Samples.QueryDynamicEntities
                     SpatialReference = SpatialReferences.Wgs84
                 };
 
+                // Get the path to the flight data JSON file.
                 string dataPath = DataManager.GetDataFolder("c78e297e99ad4572a48cdcd0b54bed30", "phx_air_traffic.json");
                 _dynamicEntityDataSource = new CustomStreamService(dataSourceInfo, dataPath, TimeSpan.FromMilliseconds(100));
 
+                // Create and configure the dynamic entity layer.
                 _dynamicEntityLayer = new DynamicEntityLayer(_dynamicEntityDataSource)
                 {
                     TrackDisplayProperties =
@@ -128,6 +130,7 @@ namespace ArcGIS.WPF.Samples.QueryDynamicEntities
                     }
                 };
 
+                // Define a simple airplane symbol for the dynamic entities.
                 var labelDefinition = new LabelDefinition(
                     new SimpleLabelExpression("[flight_number]"),
                     new TextSymbol
@@ -141,10 +144,12 @@ namespace ArcGIS.WPF.Samples.QueryDynamicEntities
                     Placement = LabelingPlacement.PointAboveCenter
                 };
 
+                // Create a simple renderer with an airplane symbol.
                 _dynamicEntityLayer.LabelDefinitions.Add(labelDefinition);
                 _dynamicEntityLayer.LabelsEnabled = true;
                 MyMapView.Map.OperationalLayers.Add(_dynamicEntityLayer);
 
+                // Load the dynamic entity layer.
                 await _dynamicEntityDataSource.ConnectAsync();
             }
             catch (Exception ex)
@@ -154,17 +159,19 @@ namespace ArcGIS.WPF.Samples.QueryDynamicEntities
             }
         }
 
-        // Handles query selection from dropdown and initiates appropriate query type
+        // Handles query selection from dropdown and initiates appropriate query type.
         private async void OnQuerySelectionChanged(object sender, SelectionChangedEventArgs e)
         {
             var selectedItem = QueryDropdown.SelectedItem as ComboBoxItem;
             if (selectedItem == null)
                 return;
 
+            // Determine the query type based on the selected item's tag.
             var queryType = selectedItem.Tag as string;
             if (string.IsNullOrEmpty(queryType))
                 return;
 
+            // If querying by track ID, show the flight number input dialog. For other query types, perform the query directly.
             if (queryType == "QueryTrackId")
             {
                 FlightNumberDialog.Visibility = Visibility.Visible;
@@ -177,7 +184,7 @@ namespace ArcGIS.WPF.Samples.QueryDynamicEntities
             }
         }
 
-        // Processes flight number input and executes track ID query
+        // Processes flight number input and executes track ID query.
         private async void OnFlightNumberQuery(object sender, RoutedEventArgs e)
         {
             var flightNumber = FlightNumberInput.Text?.Trim();
@@ -193,7 +200,7 @@ namespace ArcGIS.WPF.Samples.QueryDynamicEntities
             FlightNumberInput.Clear();
         }
 
-        // Cancels flight number input and returns to query selection
+        // Cancels flight number input and returns to query selection.
         private void OnFlightNumberCancel(object sender, RoutedEventArgs e)
         {
             FlightNumberDialog.Visibility = Visibility.Collapsed;
@@ -201,7 +208,7 @@ namespace ArcGIS.WPF.Samples.QueryDynamicEntities
             QueryDropdown.SelectedIndex = -1;
         }
 
-        // Executes the specified query type against the dynamic entity data source
+        // Executes the specified query type against the dynamic entity data source.
         private async Task PerformQuery(string queryType, string flightNumber = null)
         {
             try
@@ -222,10 +229,12 @@ namespace ArcGIS.WPF.Samples.QueryDynamicEntities
                     return;
                 }
 
+                // Set up query parameters based on the selected query type.
                 var queryParameters = new DynamicEntityQueryParameters();
 
                 switch (queryType)
                 {
+                    // Query for entities within the Phoenix airport buffer.
                     case "QueryGeometry":
                         queryParameters.Geometry = _phoenixAirportBuffer;
                         queryParameters.SpatialRelationship = SpatialRelationship.Intersects;
@@ -233,11 +242,13 @@ namespace ArcGIS.WPF.Samples.QueryDynamicEntities
                         ResultsDescription.Text = "Flights within 15 miles of PHX";
                         break;
 
+                    // Query for entities with specific attributes (status and arrival airport).
                     case "QueryAttributes":
                         queryParameters.WhereClause = "status = 'In flight' AND arrival_airport = 'PHX'";
                         ResultsDescription.Text = "Flights arriving in PHX";
                         break;
 
+                    // Query for a specific entity by its track ID (flight number).
                     case "QueryTrackId":
                         if (!string.IsNullOrEmpty(flightNumber))
                         {
@@ -252,6 +263,7 @@ namespace ArcGIS.WPF.Samples.QueryDynamicEntities
 
                 IEnumerable<DynamicEntity> results = null;
 
+                // Execute the query against the data source.
                 if (_dynamicEntityLayer?.DataSource != null)
                 {
                     results = await _dynamicEntityLayer.DataSource.QueryDynamicEntitiesAsync(queryParameters);
@@ -259,6 +271,7 @@ namespace ArcGIS.WPF.Samples.QueryDynamicEntities
 
                 results = results ?? Enumerable.Empty<DynamicEntity>();
 
+                // Process and display the query results.
                 if (results.Any())
                 {
                     foreach (var entity in results)
@@ -292,7 +305,7 @@ namespace ArcGIS.WPF.Samples.QueryDynamicEntities
             }
         }
 
-        // Centers the map view on the selected flight's current position
+        // Centers the map view on the selected flight's current position.
         private void OnResultSelectionChanged(object sender, SelectionChangedEventArgs e)
         {
             if (ResultsList.SelectedItem is FlightInfo flight && flight.Entity != null)
@@ -305,7 +318,7 @@ namespace ArcGIS.WPF.Samples.QueryDynamicEntities
             }
         }
 
-        // Closes the results panel and resets the UI to the initial state
+        // Closes the results panel and resets the UI to the initial state.
         private void OnCloseResults(object sender, RoutedEventArgs e)
         {
             ResultsPanel.Visibility = Visibility.Collapsed;
@@ -319,223 +332,6 @@ namespace ArcGIS.WPF.Samples.QueryDynamicEntities
 
             _bufferGraphicsOverlay.IsVisible = false;
             QueryDropdown.SelectedIndex = -1;
-        }
-
-        private class CustomStreamService : DynamicEntityDataSource
-        {
-            private readonly DynamicEntityDataSourceInfo _info;
-            private readonly string _dataPath;
-            private readonly TimeSpan _delay;
-            private CancellationTokenSource _cancellationTokenSource;
-            private Task _dataFeedTask;
-
-            // Initializes the custom stream service with data source info and file path
-            public CustomStreamService(DynamicEntityDataSourceInfo info, string dataPath, TimeSpan delay)
-            {
-                _info = info;
-                _dataPath = dataPath;
-                _delay = delay;
-            }
-
-            // Returns the data source information when loading the service
-            protected override Task<DynamicEntityDataSourceInfo> OnLoadAsync() =>
-                Task.FromResult(_info);
-
-            // Starts the data feed processing task when connecting to the service
-            protected override Task OnConnectAsync(CancellationToken cancellationToken)
-            {
-                _cancellationTokenSource = new CancellationTokenSource();
-                _dataFeedTask = ProcessDataFeedAsync(_cancellationTokenSource.Token);
-                return Task.CompletedTask;
-            }
-
-            // Cancels the data feed task and cleans up resources when disconnecting
-            protected override async Task OnDisconnectAsync()
-            {
-                _cancellationTokenSource?.Cancel();
-
-                if (_dataFeedTask != null)
-                {
-                    try
-                    {
-                        await _dataFeedTask;
-                    }
-                    catch (OperationCanceledException)
-                    {
-                    }
-                }
-
-                _cancellationTokenSource?.Dispose();
-                _cancellationTokenSource = null;
-            }
-
-            // Reads and processes flight data from JSON file line by line with delays
-            private async Task ProcessDataFeedAsync(CancellationToken cancellationToken)
-            {
-                try
-                {
-                    if (!File.Exists(_dataPath))
-                        throw new FileNotFoundException("Flight data file not found.", _dataPath);
-
-                    var lines = await Task.Run(() => File.ReadAllLines(_dataPath), cancellationToken);
-
-                    foreach (var line in lines)
-                    {
-                        if (cancellationToken.IsCancellationRequested) break;
-
-                        try
-                        {
-                            using (JsonDocument document = JsonDocument.Parse(line))
-                            {
-                                var root = document.RootElement;
-
-                                if (root.TryGetProperty("geometry", out JsonElement geometryElement))
-                                {
-                                    double x = 0, y = 0;
-
-                                    if (geometryElement.TryGetProperty("x", out JsonElement xElement))
-                                        x = xElement.GetDouble();
-
-                                    if (geometryElement.TryGetProperty("y", out JsonElement yElement))
-                                        y = yElement.GetDouble();
-
-                                    var point = new MapPoint(x, y, SpatialReferences.Wgs84);
-                                    var attributes = new Dictionary<string, object>();
-
-                                    if (root.TryGetProperty("attributes", out JsonElement attributesElement))
-                                    {
-                                        foreach (var property in attributesElement.EnumerateObject())
-                                        {
-                                            object value = null;
-
-                                            if (property.Value.ValueKind == JsonValueKind.String)
-                                            {
-                                                value = property.Value.GetString();
-                                            }
-                                            else if (property.Value.ValueKind == JsonValueKind.Number)
-                                            {
-                                                if (property.Value.TryGetInt32(out int intValue))
-                                                    value = intValue;
-                                                else
-                                                    value = property.Value.GetDouble();
-                                            }
-                                            else if (property.Value.ValueKind == JsonValueKind.True)
-                                            {
-                                                value = true;
-                                            }
-                                            else if (property.Value.ValueKind == JsonValueKind.False)
-                                            {
-                                                value = false;
-                                            }
-                                            else if (property.Value.ValueKind == JsonValueKind.Null)
-                                            {
-                                                value = null;
-                                            }
-                                            else
-                                            {
-                                                value = property.Value.ToString();
-                                            }
-
-                                            attributes[property.Name] = value;
-                                        }
-                                    }
-                                    AddObservation(point, attributes);
-                                }
-                            }
-                            await Task.Delay(_delay, cancellationToken);
-                        }
-                        catch (JsonException ex)
-                        {
-                            System.Diagnostics.Debug.WriteLine($"Error parsing JSON line: {ex.Message}");
-                            continue;
-                        }
-                    }
-                }
-                catch (Exception ex)
-                {
-                    System.Diagnostics.Debug.WriteLine($"Data feed error: {ex.Message}");
-                }
-            }
-        }
-
-        private class FlightInfo : INotifyPropertyChanged
-        {
-            public DynamicEntity Entity { get; }
-            private string _flightNumber;
-            private string _aircraft;
-            private string _altitude;
-            private string _speed;
-            private string _heading;
-            private string _status;
-            private string _arrivalAirport;
-
-            // Initializes flight info with entity and subscribes to change events
-            public FlightInfo(DynamicEntity entity)
-            {
-                Entity = entity;
-                UpdateAttributes();
-                Entity.DynamicEntityChanged += OnEntityChanged;
-            }
-
-            // Initializes flight info with entity and subscribes to change events
-            private bool SetField<T>(ref T field, T value, [CallerMemberName] string name = null)
-            {
-                if (EqualityComparer<T>.Default.Equals(field, value)) return false;
-                field = value;
-                OnPropertyChanged(name);
-                return true;
-            }
-
-            public string FlightNumber { get => _flightNumber; private set => SetField(ref _flightNumber, value); }
-            public string Aircraft { get => _aircraft; private set => SetField(ref _aircraft, value); }
-            public string Altitude { get => _altitude; private set => SetField(ref _altitude, value); }
-            public string Speed { get => _speed; private set => SetField(ref _speed, value); }
-            public string Heading { get => _heading; private set => SetField(ref _heading, value); }
-            public string Status { get => _status; private set => SetField(ref _status, value); }
-            public string ArrivalAirport { get => _arrivalAirport; private set => SetField(ref _arrivalAirport, value); }
-
-            // Retrieves attribute value from entity or returns default if not found
-            private string GetAttribute(string key, string defaultValue)
-            {
-                if (Entity?.Attributes != null && Entity.Attributes.TryGetValue(key, out var value))
-                    return value?.ToString() ?? defaultValue;
-                return defaultValue;
-            }
-
-            // Formats numeric string to zero decimal places or returns original if not numeric
-            private string FormatNumber(string value)
-            {
-                if (double.TryParse(value, out double number))
-                    return number.ToString("F0");
-                return value;
-            }
-
-            // Formats numeric string to zero decimal places or returns original if not numeric
-            private void UpdateAttributes()
-            {
-                FlightNumber = GetAttribute("flight_number", "N/A");
-                Aircraft = GetAttribute("aircraft", "Unknown");
-                Altitude = FormatNumber(GetAttribute("altitude_feet", "0"));
-                Speed = FormatNumber(GetAttribute("speed", "0"));
-                Heading = FormatNumber(GetAttribute("heading", "0"));
-                Status = GetAttribute("status", "Unknown");
-                ArrivalAirport = GetAttribute("arrival_airport", "N/A");
-            }
-
-            // Formats numeric string to zero decimal places or returns original if not numeric
-            private void OnEntityChanged(object sender, DynamicEntityChangedEventArgs e)
-            {
-                if (e.ReceivedObservation != null)
-                {
-                    UpdateAttributes();
-                }
-            }
-
-            public event PropertyChangedEventHandler PropertyChanged;
-
-            // Formats numeric string to zero decimal places or returns original if not numeric
-            protected virtual void OnPropertyChanged([CallerMemberName] string propertyName = null) =>
-                   PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
         }
     }
 }
